compose_send_menu_popover: Rename module to reflect broader uses.
