+++ tools/test-js-with-node
@@ -75,6 +75,7 @@ EXEMPT_FILES = make_set(
         "web/src/compose_popovers.js",
         "web/src/compose_recipient.ts",
         "web/src/compose_reply.js",
+        "web/src/compose_send_menu_popover.js",
         "web/src/compose_setup.js",
         "web/src/compose_state.ts",
         "web/src/compose_textarea.ts",
@@ -192,7 +193,6 @@ EXEMPT_FILES = make_set(
         "web/src/scheduled_messages.ts",
         "web/src/scheduled_messages_feed_ui.ts",
         "web/src/scheduled_messages_overlay_ui.js",
-        "web/src/scheduled_messages_popover.js",
         "web/src/scheduled_messages_ui.js",
         "web/src/scroll_bar.ts",
         "web/src/scroll_util.ts",
