+++ web/src/ui_init.js
@@ -27,6 +27,7 @@ import * as compose_pm_pill from "./compose_pm_pill";
 import * as compose_popovers from "./compose_popovers";
 import * as compose_recipient from "./compose_recipient";
 import * as compose_reply from "./compose_reply";
+import * as compose_send_menu_popover from "./compose_send_menu_popover";
 import * as compose_setup from "./compose_setup";
 import * as compose_textarea from "./compose_textarea";
 import * as compose_tooltips from "./compose_tooltips";
@@ -98,7 +99,6 @@ import * as reload_setup from "./reload_setup";
 import * as resize_handler from "./resize_handler";
 import * as scheduled_messages from "./scheduled_messages";
 import * as scheduled_messages_overlay_ui from "./scheduled_messages_overlay_ui";
-import * as scheduled_messages_popover from "./scheduled_messages_popover";
 import * as scheduled_messages_ui from "./scheduled_messages_ui";
 import * as scroll_bar from "./scroll_bar";
 import * as scroll_util from "./scroll_util";
@@ -682,7 +682,7 @@ export function initialize_everything(state_data) {
     user_topic_popover.initialize();
     topic_popover.initialize();
     message_actions_popover.initialize();
-    scheduled_messages_popover.initialize();
+    compose_send_menu_popover.initialize();
 
     realm_user_settings_defaults.initialize(realm_settings_defaults_params);
     people.initialize(current_user.user_id, people_params);
